+++ tests/Fixtures/TestBundle/Entity/Dummy.php
@@ -218,7 +218,7 @@ class Dummy
         return $this->description;
     }
 
-    public function hasRole($role)
+    public function fooBar($baz)
     {
     }